Move downtime notification actions and reducer into directory
