+++ src/js/common/reducers/index.js
@@ -1,3 +1,3 @@
-import scheduledDowntime from './downtime';
+import scheduledDowntime from '../../../platform/monitoring/DowntimeNotification/reducer';
 
 export { scheduledDowntime };
